requirements.txt
tetgen
numpy
numpy-stl
tk
matplotlib
scipy
reportlab
Pillow

